Split nbd ioctl handling into a separate "sys" module
This will make it easier to separate attaching an NBD device from
serving NBD protocol.
