+++ src/lib.rs
@@ -1,4 +1,5 @@
 mod device;
 mod nbd;
+mod sys;
 
 pub use device::{attach_device, BlockDevice};
